:arrow_right: :alarm_clock: :athletic_shoe: 18130번 해결중
형변환 안하고 int * int 를 했더니 오버플로우 발생... 해당 부분은 명시적 형변환(long)으로 해결함
해결하고 나니 채점 19%에서 시간초과가 뜸... 중간에 while문을 써서 그런가

+++ questions/src/main/java/baekjoon/processing/Main18130.java
@@ -26,7 +26,7 @@ public class Main18130 {
             int timeflow = fanInfo[1], section = 1; //추가비용 첫 구간으로 초기화
             while(timeflow < time) {
                 //추가비용 부과
-                expense += section*fanInfo[2];
+                expense += (long) section*fanInfo[2];
                 //다음 추가비용 부과구간
                 timeflow += fanInfo[1];
                 section++;
